+++ foo_ui_columns/playlist_switcher_inline_edit.cpp
@@ -4,11 +4,11 @@
 namespace cui::panels::playlist_switcher {
 
 bool PlaylistSwitcher::notify_before_create_inline_edit(
-    const pfc::list_base_const_t<t_size>& indices, unsigned column, bool b_source_mouse)
+    const pfc::list_base_const_t<t_size>& indices, size_t column, bool b_source_mouse)
 {
     return column == 0 && indices.get_count() == 1;
 }
-bool PlaylistSwitcher::notify_create_inline_edit(const pfc::list_base_const_t<t_size>& indices, unsigned column,
+bool PlaylistSwitcher::notify_create_inline_edit(const pfc::list_base_const_t<t_size>& indices, size_t column,
     pfc::string_base& p_text, t_size& p_flags, mmh::ComPtr<IUnknown>& pAutocompleteEntries)
 {
     t_size indices_count = indices.get_count();
